% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{enrich_dbs}
\alias{enrich_dbs}
\title{function to call enrich fro pathways}
\usage{
enrich_dbs(gene_list, dbs, adjpval_filter = 0.05)
}
\arguments{
\item{gene_list}{Character vector of gene names or data.}

\item{dbs}{Character vector of databases to search. See https://maayanlab.cloud/Enrichr/ for available databases}

\item{adjpval_filter}{=0.05}
}
\value{
a dataframe
}
\description{
function to call enrich fro pathways
}
\details{
This is some descriptio of this function.

function to call enrich fro pathways
}
