% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_med_expname.R
\name{find_med_expname}
\alias{find_med_expname}
\title{Auto Find Mediation Analysis}
\usage{
find_med_expname(
  exposure_name,
  pop = "European",
  eqtl = FALSE,
  num_exp = 1000,
  num_out = 1000,
  pval = 0.05,
  renew_id = FALSE,
  save_path = "D:\\\\课题组的活\\\\开题\\\\1"
)
}
\arguments{
\item{pop}{Population to consider (default is "European").}

\item{eqtl}{Logical flag to include eQTL analysis (default is FALSE).}

\item{num_exp}{Number of exposures to process (default is 30).}

\item{num_out}{Number of outcomes to process (default is 30).}

\item{pval}{P-value threshold (default is 0.05).}

\item{renew_id}{Logical flag to renew available outcomes ID (default is FALSE).}

\item{save_path}{Path to save results (default is "D:\\课题组的活\\开题\\1").}

\item{trait_name}{Name of the trait to analyze.}

\item{trait_id}{ID of the trait to analyze.}

\item{auto}{Logical flag to perform automated analysis (default is FALSE).}
}
\value{
NULL. Writes results to a file.
}
\description{
This function performs mediation analysis for a specified trait.
}
\examples{
auto_find_med(trait_name="example_trait", trait_id="example_id")
}
