% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_risks_method.R
\name{find_risks_method}
\alias{find_risks_method}
\title{Auto Find Mediation Analysis}
\usage{
find_risks_method(
  outcome_id = "ukb-d-KNEE_ARTHROSIS",
  num1 = 1,
  num2 = 42334,
  pval = 5e-08,
  clump_r2 = 0.001,
  clump_kb = 10000,
  proxies = TRUE,
  pop = "European",
  renew_id = FALSE,
  action = 3,
  methods = c("IVW", "MR-Egger", "MR-PRESSO", "Weighted Median", "Simple Mode",
    "Weighted Mode"),
  save_format = "rds",
  save_path = "D:\\\\课题组的活\\\\开题\\\\1"
)
}
\arguments{
\item{pval}{P-value threshold (default is 0.05).}

\item{pop}{Population to consider (default is "European").}

\item{renew_id}{Logical flag to renew available outcomes ID (default is FALSE).}

\item{save_path}{Path to save results (default is "D:\\课题组的活\\开题\\1").}

\item{trait_name}{Name of the trait to analyze.}

\item{trait_id}{ID of the trait to analyze.}

\item{eqtl}{Logical flag to include eQTL analysis (default is FALSE).}

\item{auto}{Logical flag to perform automated analysis (default is FALSE).}

\item{num_exp}{Number of exposures to process (default is 30).}

\item{num_out}{Number of outcomes to process (default is 30).}
}
\value{
NULL. Writes results to a file.
}
\description{
This function performs mediation analysis for a specified trait.
}
\examples{
auto_find_med(trait_name="example_trait", trait_id="example_id")
}
