% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_eQTL_for_snp}
\alias{get_eQTL_for_snp}
\title{function to call enrich fro pathways}
\usage{
get_eQTL_for_snp(variant)
}
\arguments{
\item{variant}{the SNP list.}
}
\value{
a dataframe
}
\description{
function to get eQTL by queried snp.
}
\details{
This is some descriptio of this function.

function to get eQTL by queried snp.
}
