% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{kable_it}
\alias{kable_it}
\title{query wrapper function}
\usage{
kable_it(df)
}
\value{
a dataframe
}
\description{
query wrapper function
}
\details{
This is some descriptio of this function.

query wrapper function
}
