% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{makeforest_plot}
\alias{makeforest_plot}
\title{query wrapper function}
\usage{
makeforest_plot(res_sub, my_title)
}
\arguments{
\item{res_sub}{mr result}

\item{my_title}{the title name of my plot}
}
\value{
a dataframe
}
\description{
make forest for query result为查询结果创建森林图
}
\details{
This is some descriptio of this function.

make forest for query result
}
