% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{map_ensembl_to_genename}
\alias{map_ensembl_to_genename}
\title{function to call enrich fro pathways}
\usage{
map_ensembl_to_genename(values, GRCh = 37)
}
\arguments{
\item{values}{Values of the filter, e.g. vector of affy IDs. If multiple filters are specified then the argument should be a list of vectors of which the position of each vector corresponds to the position of the filters in the filters argument.}

\item{GRCh}{=37 or 38.}
}
\value{
a dataframe
}
\description{
function to get eQTL by queried snp.
}
\details{
This is some descriptio of this function.

function to get eQTL by queried snp.
}
