% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{mr_eve}
\alias{mr_eve}
\title{function to query mr result based on gived outcome_id in epigraphdb database website.}
\usage{
mr_eve(
  outcome_id = "ukb-a-100",
  outcome_trait,
  p1 = 1,
  p2 = 5e-08,
  save_path = "D:\\\\课题组的活\\\\开题\\\\1"
)
}
\arguments{
\item{outcome_id}{the mabase id of outcome.}

\item{outcome_trait}{the outcome trait.}

\item{p1}{=1, query all MR results for the outcomes, not restricting by p-value.}

\item{p2}{=5e-8, V2 - alternative - more stringent with pval < 5e-8 for all.}

\item{save_path}{save the result to file path.}
}
\value{
a dataframe
}
\description{
function to query mr result based on gived outcome_id in epigraphdb database website.
}
\details{
This is some descriptio of this function.

function to query mr result based on gived outcome_id in epigraphdb database website.
}
