% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot__bubble_chart.R
\name{plot_bubble_chart}
\alias{plot_bubble_chart}
\title{绘制气泡图}
\usage{
plot_bubble_chart(
  results,
  p_column = "mr.pval",
  x_column = "mr.b",
  title = "title",
  x_lab = "Beta",
  y_lab = "-log10(P值)",
  labtext = "Beta绝对值",
  name = "exposure.trait",
  col_func_negative = colorRampPalette(c("lightblue1", "deepskyblue4")),
  col_func_positive = colorRampPalette(c("pink", "firebrick4")),
  width = 40,
  height = 20,
  dpi = 300,
  output_file = NULL,
  output_format = "png"
)
}
\arguments{
\item{results}{数据框，包含 MR 分析结果}

\item{p_column}{字符串，表示 P 值所在的列名，默认 "mr.pval"}

\item{x_column}{字符串，表示 Beta 值所在的列名，默认 "mr.b"}

\item{title}{图表标题}

\item{x_lab}{x 轴标签，默认 "Beta"}

\item{y_lab}{y 轴标签，默认 "-log10(P值)"}

\item{labtext}{图例中显示的大小说明，默认 "Beta绝对值"}

\item{name}{用于显示标签的列名}

\item{col_func_negative}{颜色映射函数（负方向）}

\item{col_func_positive}{颜色映射函数（正方向）}

\item{width}{图宽（单位：cm）}

\item{height}{图高（单位：cm）}

\item{dpi}{分辨率}

\item{output_file}{输出文件路径，默认为 NULL 不输出}

\item{output_format}{输出格式："png"、"jpeg" 或 "pdf"}
}
\value{
无返回值，绘图并打印或保存
}
\description{
根据 MR 分析结果绘制气泡图，展示 Beta 值与 P 值的关系，并通过颜色和大小标记显著性和方向。
}
