% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{query_epigraphdb_as_table}
\alias{query_epigraphdb_as_table}
\title{query wrapper function 查询包装器功能}
\usage{
query_epigraphdb_as_table(query)
}
\arguments{
\item{query}{A list of parameters associated with the query endpoint.
与查询端点相关联的参数列表。}
}
\value{
a dataframe
}
\description{
query wrapper function查询包装器功能
}
\details{
This is some descriptio of this function.这是这个函数的一些描述。

This is a general purpose function to send data request which can be used when there has not been an R equivalent package function to an API endpoint.
}
