% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{run_local_mr_all_outcomes}
\alias{run_local_mr_all_outcomes}
\title{query wrapper function}
\usage{
run_local_mr_all_outcomes(trait_investigated, outcomes_to_try)
}
\arguments{
\item{trait_investigated}{the trait id}

\item{outcomes_to_try}{the outcome id}
}
\value{
a dataframe
}
\description{
run local mr all outcomes
}
\details{
This is some descriptio of this function.

run local mr all outcomes
}
